pfringcapturethread.cpp: instead of calculating a wrong packet length, use header->tp_len and header->tp_snaplen as appropriate.
The packet length was previously calculated using (header->tp_len -
pkt_offset), which did not make any sense.

As explained in a previous commit, header->tp_len is the length of the
packet as received. Subtracting some random PF_PACKET-specific header
length will not lead to any valid packet length we could use.

Instead, header->tp_snaplen is a little bit more interesting to us,
because this is the captured packet length, which give-or-take tells us
the actual amount of bytes written to the data buffer.

Note the "give-or-take" there - this can be true, but need not be. A
follow-up commit will fix and expand on this.

This also allows us to drop the pkt_offset variable, which by itself was
not wrong (giving us the offset from start of the header to the optional
MAC header), but certainly implementation-specific.