pfringcapturethread.cpp: make sure that we not read data past a frame.
A frame is the maximum amount of data a captured packet can contain -
and this includes the PF_PACKET RX ring buffer headers.

Reading past a frame will, in the best case, just add pseudo-random
garbage to the end of a packet capture (mostly the PF_PACKET RX ring
buffer header of a different capture packet), and in the worst case lead
to crashes when reading past the mmapped region.

I have experienced those type of crashes whenever the last packet in the
block was processed during the packet copy operation.

Interestingly, the crash only seems to happen when using AVX-intrisics
to copy the data (and hence when using the default optimizations), but
not when optimizations are disabled.

Regardless, reading past the region is an error and the crashes were
well-deserved.